// NetscriptDefinitions.d.ts
interface HP {
  current: number;
  max: number;
}

/** @public */
interface Skills {
  hacking: number;
  strength: number;
  defense: number;
  dexterity: number;
  agility: number;
  charisma: number;
  intelligence: number;
}

// TODO: provide same treatment to CodingContractData as for SleeveTask (actual types)
/**
 * Coding contract data will differ depending on coding contract.
 * @public
 */
type CodingContractData = any;

/** @public */
type ScriptArg = string | number | boolean;

/** @public */
type FilenameOrPID = number | string;

/** @public */
interface Person {
  hp: HP;
  skills: Skills;
  exp: Skills;
  mults: Multipliers;
  city: CityName;
}

/** @public */
interface Player extends Person {
  money: number;
  numPeopleKilled: number;
  entropy: number;
  jobs: Partial<Record<CompanyName, JobName>>;
  factions: string[];
  totalPlaytime: number;
  location: LocationName;
  karma: number;
}

/** @public */
interface SleevePerson extends Person {
  /** Number 0-100 Experience earned and shared is multiplied with shock% before sync% */
  shock: number;
  /** Number 1-100 Experience earned by this sleeve and shared with the player is multiplied with sync% after shock% */
  sync: number;
  /** Number 1-100 initial Value of sync on BN start */
  memory: number;
  /** Number of 200ms cycles which are stored as bonus time */
  storedCycles: number;
}
// ...
interface Multipliers {
  /** Multiplier to hacking skill */
  hacking: number;
  /** Multiplier to strength skill */
  strength: number;
  /** Multiplier to defense skill */
  defense: number;
  /** Multiplier to dexterity skill */
  dexterity: number;
  /** Multiplier to agility skill */
  agility: number;
  /** Multiplier to charisma skill */
  charisma: number;
  /** Multiplier to hacking experience gain rate */
  hacking_exp: number;
  /** Multiplier to strength experience gain rate */
  strength_exp: number;
  /** Multiplier to defense experience gain rate */
  defense_exp: number;
  /** Multiplier to dexterity experience gain rate */
  dexterity_exp: number;
  /** Multiplier to agility experience gain rate */
  agility_exp: number;
  /** Multiplier to charisma experience gain rate */
  charisma_exp: number;
  /** Multiplier to chance of successfully performing a hack */
  hacking_chance: number;
  /** Multiplier to hacking speed */
  hacking_speed: number;
  /** Multiplier to amount of money the player gains from hacking */
  hacking_money: number;
  /** Multiplier to amount of money injected into servers using grow */
  hacking_grow: number;
  /** Multiplier to amount of reputation gained when working */
  company_rep: number;
  /** Multiplier to amount of reputation gained when working */
  faction_rep: number;
  /** Multiplier to amount of money gained from crimes */
  crime_money: number;
  /** Multiplier to crime success rate */
  crime_success: number;
  /** Multiplier to amount of money gained from working */
  work_money: number;
  /** Multiplier to amount of money produced by Hacknet Nodes */
  hacknet_node_money: number;
  /** Multiplier to cost of purchasing a Hacknet Node */
  hacknet_node_purchase_cost: number;
  /** Multiplier to cost of ram for a Hacknet Node */
  hacknet_node_ram_cost: number;
  /** Multiplier to cost of core for a Hacknet Node */
  hacknet_node_core_cost: number;
  /** Multiplier to cost of leveling up a Hacknet Node */
  hacknet_node_level_cost: number;
  /** Multiplier to Bladeburner max stamina */
  bladeburner_max_stamina: number;
  /** Multiplier to Bladeburner stamina gain rate */
  bladeburner_stamina_gain: number;
  /** Multiplier to effectiveness in Bladeburner Field Analysis */
  bladeburner_analysis: number;
  /** Multiplier to success chance in Bladeburner contracts/operations */
  bladeburner_success_chance: number;
}
// ...
type SleeveBladeburnerTask = {
  type: "BLADEBURNER";
  actionType: "General" | "Contracts";
  actionName: string;
  cyclesWorked: number;
  cyclesNeeded: number;
  nextCompletion: Promise<void>;
  tasksCompleted: number;
};

/** @public */
type SleeveClassTask = {
  type: "CLASS";
  classType: UniversityClassType | GymType | `${UniversityClassType}` | `${GymType}`;
  location: LocationName | `${LocationName}`;
};

/** @public */
type SleeveCompanyTask = { type: "COMPANY"; companyName: CompanyName };

/** @public */
type SleeveCrimeTask = {
  type: "CRIME";
  crimeType: CrimeType | `${CrimeType}`;
  cyclesWorked: number;
  cyclesNeeded: number;
  tasksCompleted: number;
};

/** @public */
type SleeveFactionTask = {
  type: "FACTION";
  factionWorkType: FactionWorkType | `${FactionWorkType}`;
  factionName: string;
};

/** @public */
type SleeveInfiltrateTask = {
  type: "INFILTRATE";
  cyclesWorked: number;
  cyclesNeeded: number;
  nextCompletion: Promise<void>;
};

/** @public */
type SleeveRecoveryTask = { type: "RECOVERY" };

/** @public */
type SleeveSupportTask = { type: "SUPPORT" };

/** @public */
type SleeveSynchroTask = { type: "SYNCHRO" };

/** Object representing a sleeve current task.
 * @public */
export type SleeveTask =
  | SleeveBladeburnerTask
  | SleeveClassTask
  | SleeveCompanyTask
  | SleeveCrimeTask
  | SleeveFactionTask
  | SleeveInfiltrateTask
  | SleeveRecoveryTask
  | SleeveSupportTask
  | SleeveSynchroTask;
// ...
declare enum CrimeType {
  shoplift = "Shoplift",
  robStore = "Rob Store",
  mug = "Mug",
  larceny = "Larceny",
  dealDrugs = "Deal Drugs",
  bondForgery = "Bond Forgery",
  traffickArms = "Traffick Arms",
  homicide = "Homicide",
  grandTheftAuto = "Grand Theft Auto",
  kidnap = "Kidnap",
  assassination = "Assassination",
  heist = "Heist",
}

/** @public */
declare enum FactionWorkType {
  hacking = "hacking",
  field = "field",
  security = "security",
}

/** @public */
declare enum UniversityClassType {
  computerScience = "Computer Science",
  dataStructures = "Data Structures",
  networks = "Networks",
  algorithms = "Algorithms",
  management = "Management",
  leadership = "Leadership",
}

/** @public */
declare enum GymType {
  strength = "str",
  defense = "def",
  dexterity = "dex",
  agility = "agi",
}

/** @public */
declare enum JobName {
  software0 = "Software Engineering Intern",
  software1 = "Junior Software Engineer",
  software2 = "Senior Software Engineer",
  software3 = "Lead Software Developer",
  software4 = "Head of Software",
  software5 = "Head of Engineering",
  software6 = "Vice President of Technology",
  software7 = "Chief Technology Officer",
  IT0 = "IT Intern",
  IT1 = "IT Analyst",
  IT2 = "IT Manager",
  IT3 = "Systems Administrator",
  securityEng = "Security Engineer",
  networkEng0 = "Network Engineer",
  networkEng1 = "Network Administrator",
  business0 = "Business Intern",
  business1 = "Business Analyst",
  business2 = "Business Manager",
  business3 = "Operations Manager",
  business4 = "Chief Financial Officer",
  business5 = "Chief Executive Officer",
  security0 = "Security Guard",
  security1 = "Security Officer",
  security2 = "Security Supervisor",
  security3 = "Head of Security",
  agent0 = "Field Agent",
  agent1 = "Secret Agent",
  agent2 = "Special Operative",
  waiter = "Waiter",
  employee = "Employee",
  softwareConsult0 = "Software Consultant",
  softwareConsult1 = "Senior Software Consultant",
  businessConsult0 = "Business Consultant",
  businessConsult1 = "Senior Business Consultant",
  waiterPT = "Part-time Waiter",
  employeePT = "Part-time Employee",
}
// ...
declare enum CityName {
  Aevum = "Aevum",
  Chongqing = "Chongqing",
  Sector12 = "Sector-12",
  NewTokyo = "New Tokyo",
  Ishima = "Ishima",
  Volhaven = "Volhaven",
}

/** Names of all locations
 * @public */
declare enum LocationName {
  AevumAeroCorp = "AeroCorp",
  AevumBachmanAndAssociates = "Bachman & Associates",
  AevumClarkeIncorporated = "Clarke Incorporated",
  AevumCrushFitnessGym = "Crush Fitness Gym",
  AevumECorp = "ECorp",
  AevumFulcrumTechnologies = "Fulcrum Technologies",
  AevumGalacticCybersystems = "Galactic Cybersystems",
  AevumNetLinkTechnologies = "NetLink Technologies",
  AevumPolice = "Aevum Police Headquarters",
  AevumRhoConstruction = "Rho Construction",
  AevumSnapFitnessGym = "Snap Fitness Gym",
  AevumSummitUniversity = "Summit University",
  AevumWatchdogSecurity = "Watchdog Security",
  AevumCasino = "Iker Molina Casino",

  ChongqingKuaiGongInternational = "KuaiGong International",
  ChongqingSolarisSpaceSystems = "Solaris Space Systems",
  ChongqingChurchOfTheMachineGod = "Church of the Machine God",

  Sector12AlphaEnterprises = "Alpha Enterprises",
  Sector12BladeIndustries = "Blade Industries",
  Sector12CIA = "Central Intelligence Agency",
  Sector12CarmichaelSecurity = "Carmichael Security",
  Sector12CityHall = "Sector-12 City Hall",
  Sector12DeltaOne = "DeltaOne",
  Sector12FoodNStuff = "FoodNStuff",
  Sector12FourSigma = "Four Sigma",
  Sector12IcarusMicrosystems = "Icarus Microsystems",
  Sector12IronGym = "Iron Gym",
  Sector12JoesGuns = "Joe's Guns",
  Sector12MegaCorp = "MegaCorp",
  Sector12NSA = "National Security Agency",
  Sector12PowerhouseGym = "Powerhouse Gym",
  Sector12RothmanUniversity = "Rothman University",
  Sector12UniversalEnergy = "Universal Energy",

  NewTokyoDefComm = "DefComm",
  NewTokyoGlobalPharmaceuticals = "Global Pharmaceuticals",
  NewTokyoNoodleBar = "Noodle Bar",
  NewTokyoVitaLife = "VitaLife",
  NewTokyoArcade = "Arcade",

  IshimaNovaMedical = "Nova Medical",
  IshimaOmegaSoftware = "Omega Software",
  IshimaStormTechnologies = "Storm Technologies",
  IshimaGlitch = "0x6C1",

  VolhavenCompuTek = "CompuTek",
  VolhavenHeliosLabs = "Helios Labs",
  VolhavenLexoCorp = "LexoCorp",
  VolhavenMilleniumFitnessGym = "Millenium Fitness Gym",
  VolhavenNWO = "NWO",
  VolhavenOmniTekIncorporated = "OmniTek Incorporated",
  VolhavenOmniaCybersystems = "Omnia Cybersystems",
  VolhavenSysCoreSecurities = "SysCore Securities",
  VolhavenZBInstituteOfTechnology = "ZB Institute of Technology",

  Hospital = "Hospital",
  Slums = "The Slums",
  TravelAgency = "Travel Agency",
  WorldStockExchange = "World Stock Exchange",

  Void = "The Void",
}

/** Names of all companies
 * @public */
declare enum CompanyName {
  ECorp = "ECorp",
  MegaCorp = "MegaCorp",
  BachmanAndAssociates = "Bachman & Associates",
  BladeIndustries = "Blade Industries",
  NWO = "NWO",
  ClarkeIncorporated = "Clarke Incorporated",
  OmniTekIncorporated = "OmniTek Incorporated",
  FourSigma = "Four Sigma",
  KuaiGongInternational = "KuaiGong International",
  FulcrumTechnologies = "Fulcrum Technologies",
  StormTechnologies = "Storm Technologies",
  DefComm = "DefComm",
  HeliosLabs = "Helios Labs",
  VitaLife = "VitaLife",
  IcarusMicrosystems = "Icarus Microsystems",
  UniversalEnergy = "Universal Energy",
  GalacticCybersystems = "Galactic Cybersystems",
  AeroCorp = "AeroCorp",
  OmniaCybersystems = "Omnia Cybersystems",
  SolarisSpaceSystems = "Solaris Space Systems",
  DeltaOne = "DeltaOne",
  GlobalPharmaceuticals = "Global Pharmaceuticals",
  NovaMedical = "Nova Medical",
  CIA = "Central Intelligence Agency",
  NSA = "National Security Agency",
  WatchdogSecurity = "Watchdog Security",
  LexoCorp = "LexoCorp",
  RhoConstruction = "Rho Construction",
  AlphaEnterprises = "Alpha Enterprises",
  Police = "Aevum Police Headquarters",
  SysCoreSecurities = "SysCore Securities",
  CompuTek = "CompuTek",
  NetLinkTechnologies = "NetLink Technologies",
  CarmichaelSecurity = "Carmichael Security",
  FoodNStuff = "FoodNStuff",
  JoesGuns = "Joe's Guns",
  OmegaSoftware = "Omega Software",
  NoodleBar = "Noodle Bar",
}